netmap: update BM pool used and free counters together
This change allows to release rxq descriptors as soon as they are
read by netmap rx_synch function, instead of waiting until the buffers are
released by user space.

This allows for performance improvement at rx side.

Note: since in current netmap implementation the number of rxq descriptors is the
same as the number of buffers in BM pool, it may happen that BM pool drop counter
increases due to lack of BM pool buffers. This behavior is normal.

Change-Id: I89dde23a2a7e3f8663037858c232402b90d94b72
Signed-off-by: Alan Winkowski <walan@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/32948
Reviewed-by: Yuval Caduri <cyuval@marvell.com>
Tested-by: Yuval Caduri <cyuval@marvell.com>