netmap: add interrupt support in netmap poll function
This code was added to enable interrupt support in mvpp2x
layer on netmap_poll() only.
By default interrupts are disabled when using netmap and
more proper way to enable it was not found.
Using nm_intr() callback to enable interrupt support in
netmap_poll() function  before calling rx_sync.
The actual enabling of interrupts will be done in rx_sync
if there is no packets received.
In case of using ioctl, the interrupts support will be
disabled.

Change-Id: Ib416d9c9dc7cbb99d68a0a4b00d58a6105997b6f
Signed-off-by: nsamsono <nsamsono@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/41329
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Alan Winkowski <walan@marvell.com>
Reviewed-by: Yuval Caduri <cyuval@marvell.com>